:truck: api: move find-finalized-sessions-with-required-action usecase from lib to src
