+++ api/src/certification/session-management/domain/usecases/find-finalized-sessions-with-required-action.js
